feat: sperate defaultLanguage to avoid circular reference
